Started on EDA
Started the EDA, I've got most of the graphs on there and the skeleton code which I plan on improving and adding to during the week. Haven't started writing about the graphs or anything else. Currently has histograms, box plots and Q-Q plots for all the numeric variables, covariance and correlation matrices and a correlogram. Feel free to add anything or lmk if I need to add anything.