Add files via upload
fixed plot colourings and figure and table referencing, and added a missing citation. I think this is worth committing to main as it is the most up-to-date version of the project report, minus photos of us (which may or may not be even relevant, come the time of the final report).